Update Group2DealershipQuery_Draft2.sql
Data input completed.  Needed to update the Prices to Decimal (8,2) as it was giving me an error message as the total numbers were exceeding our values including the decimals.

Everything loaded when executing.  I recommend adding more comments on each table on what it is.  She is looking for who did what in the comments.